fix<BE> : Favorite List 자료형 Long으로 수정

See merge request s07-webmobile1-sub2/S07P12B204!101

+++ backend/src/main/java/com/ssafy/cheertogether/member/dto/Oauth2JoinRequest.java
@@ -11,6 +11,6 @@ public class Oauth2JoinRequest {
 	private String profileImage;
 	private String role;
 	private String myInfo;
-	private List<Integer> favoriteLeagueList;
-	private List<Integer> favoriteTeamList;
+	private List<Long> favoriteLeagueList;
+	private List<Long> favoriteTeamList;
 }
